requirements.txt
streamlit==1.31.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.24.0
pycocotools>=2.0.6
matplotlib>=3.7.0
pandas>=2.0.0
streamlit-drawable-canvas>=0.9.0